assignment-3/app/garden-service/mqtt-component/requirements.txt
redis
paho-mqtt